+++ frontend/src/concepts/connectionTypes/__tests__/utils.spec.ts
@@ -17,7 +17,6 @@ import {
   getModelServingCompatibility,
   isModelServingCompatible,
   isValidEnvVar,
-  isValueFromEnvVar,
   ModelServingCompatibleTypes,
   toConnectionTypeConfigMap,
   toConnectionTypeConfigMapObj,
@@ -543,30 +542,5 @@ describe('useTrimInputHandlers', () => {
 
     trimInputOnPaste('', handleChange)(mockEvent);
     expect(handleChange).toHaveBeenCalledWith('foo');
-  });
-});
-
-describe('isValueFrom', () => {
-  it('should return true if the value is from a valueFrom envVar', () => {
-    expect(
-      isValueFromEnvVar({
-        valueFrom: {
-          secretKeyRef: {
-            name: 'test',
-            key: '',
-          },
-        },
-        name: '',
-      }),
-    ).toBe(true);
-  });
-
-  it('should return false if the value is not from a valueFrom envVar', () => {
-    expect(
-      isValueFromEnvVar({
-        value: 'test',
-        name: '',
-      }),
-    ).toBe(false);
   });
 });
